basic API working
