+++ backend/models/__init__.py
@@ -1,0 +1,3 @@
+from database import Base
+from .user_model import User
+from .note_model import Note
