+++ backend/models/note_model.py
@@ -3,17 +3,6 @@ from sqlalchemy.orm import relationship
 from database import Base
 from datetime import datetime
 
-# Model użytkownika
-class User(Base):
-    __tablename__ = 'users'
-
-    id = Column(Integer, primary_key=True, index=True)
-    email = Column(String, unique=True, index=True, nullable=False)
-    password = Column(String, nullable=False)
-    
-    # Relacja do notatek
-    notes = relationship("Note", back_populates="owner")
-
 # Model notatki
 class Note(Base):
     __tablename__ = 'notes'
@@ -27,4 +16,3 @@ class Note(Base):
 
     # Relacja z użytkownikiem
     owner = relationship("User", back_populates="notes")
-
